requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
alembic==1.4.2            # via -r requirements/prod.in
appdirs==1.4.4            # via black, virtualenv
asyncpg==0.20.1           # via databases
attrs==19.3.0             # via black, pytest
bcrypt==3.1.7             # via passlib
black==19.10b0            # via -r requirements/dev.in
certifi==2020.6.20        # via requests
cffi==1.14.0              # via bcrypt, cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via black, uvicorn
cryptography==2.9.2       # via python-jose
databases[postgresql]==0.3.2  # via -r requirements/prod.in
distlib==0.3.0            # via virtualenv
ecdsa==0.15               # via python-jose
environs==8.0.0           # via -r requirements/dev.in
factory-boy==2.12.0       # via -r requirements/dev.in
faker==4.1.1              # via factory-boy
fastapi==0.59.0           # via -r requirements/prod.in
filelock==3.0.12          # via tox, virtualenv
flake8-blind-except==0.1.1  # via -r requirements/dev.in
flake8-debugger==3.2.1    # via -r requirements/dev.in
flake8-docstrings==1.5.0  # via -r requirements/dev.in
flake8-isort==3.0.0       # via -r requirements/dev.in
flake8-polyfill==1.0.2    # via pep8-naming
flake8==3.8.3             # via -r requirements/dev.in, flake8-debugger, flake8-docstrings, flake8-isort, flake8-polyfill
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
idna==2.9                 # via requests
isort[pyproject]==4.3.21  # via -r requirements/dev.in, flake8-isort
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via mako
marshmallow==3.6.1        # via environs
mccabe==0.6.1             # via flake8
more-itertools==8.4.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.781               # via -r requirements/dev.in
orjson==3.2.1             # via -r requirements/prod.in
packaging==20.4           # via pytest, tox
passlib[bcrypt]==1.7.2    # via -r requirements/prod.in
pathspec==0.8.0           # via black
pep8-naming==0.11.1       # via -r requirements/dev.in
pluggy==0.13.1            # via pytest, tox
psycopg2-binary==2.8.5    # via databases
py==1.8.2                 # via pytest, tox
pyasn1==0.4.8             # via python-jose, rsa
pycodestyle==2.6.0        # via flake8, flake8-debugger
pycparser==2.20           # via cffi
pydantic==1.5.1           # via fastapi
pydocstyle==5.0.2         # via flake8-docstrings
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.14.0    # via -r requirements/dev.in
pytest==5.4.3             # via -r requirements/dev.in, pytest-asyncio
python-dateutil==2.8.1    # via alembic, faker
python-dotenv==0.13.0     # via environs
python-editor==1.0.4      # via alembic
python-jose[cryptography]==3.1.0  # via -r requirements/prod.in
python-multipart==0.0.5   # via -r requirements/prod.in
regex==2020.6.8           # via black
requests==2.24.0          # via -r requirements/dev.in
rsa==4.6                  # via python-jose
six==1.15.0               # via bcrypt, cryptography, ecdsa, packaging, python-dateutil, python-jose, python-multipart, tox, virtualenv
snowballstemmer==2.0.0    # via pydocstyle
sqlalchemy==1.3.17        # via -r requirements/prod.in, alembic, databases
starlette==0.13.4         # via fastapi
testfixtures==6.14.1      # via flake8-isort
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, isort, tox
tox==3.15.2               # via -r requirements/dev.in
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
urllib3==1.25.9           # via requests
uvicorn==0.11.5           # via -r requirements/prod.in
uvloop==0.14.0            # via uvicorn
virtualenv==20.0.23       # via tox
wcwidth==0.2.4            # via pytest
websockets==8.1           # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
